+++ src/components/Challenge/TodayChallengeList.tsx
@@ -9,7 +9,7 @@ const dummyChallenges: ChallengeType[] = [
   {
     id: "1",
     title: "텀블러로 커피 마시는 멋진 나는..",
-    desciption: "내용",
+    description: "내용",
     categories: ["001a", "텀블러_챌린지"],
     prover_cnt: 10,
     point: 100,
@@ -25,7 +25,7 @@ const dummyChallenges: ChallengeType[] = [
   {
     id: "1",
     title: "텀블러로 커피 마시는 멋진 나는..",
-    desciption: "내용",
+    description: "내용",
     categories: ["001a", "텀블러_챌린지"],
     prover_cnt: 10,
     point: 100,
@@ -41,7 +41,7 @@ const dummyChallenges: ChallengeType[] = [
   {
     id: "1",
     title: "텀블러로 커피 마시는 멋진 나는..",
-    desciption: "내용",
+    description: "내용",
     categories: ["001a", "텀블러_챌린지"],
     prover_cnt: 10,
     point: 100,
