+++ src/components/Peace/Post/dummyCommemts.ts
@@ -1,0 +1,32 @@
+export const dummyComments = {
+  count: 2,
+  next: "",
+  previous: "",
+  results: [
+    {
+      id: 1,
+      article_id: "",
+      writer: {
+        id: 1,
+        username: "SoyE",
+        avatar: "images/man.png",
+      },
+      content: "와! 고생하셨어요!",
+      created_at: "2022-08-30T01:10:53.265788+09:00",
+    },
+    {
+      id: 2,
+      article_id: "",
+      writer: {
+        id: 1,
+        username: "kwonsoye",
+        avatar: "images/man.png",
+      },
+      content:
+        "각급 선거관리위원회는 선거인명부의 작성등 선거사무와 국민투표사무에 관하여 관계 행정기관에 필요한 지시를 할 수 있다. 헌법재판소는 법관의 자격을 가진 9인의 재판관으로 구성하며, 재판관은 대통령이 임명한다.",
+      created_at: "2022-08-30T01:10:53.265788+09:00",
+    },
+  ],
+};
+
+export default dummyComments;
