--- src/components/Peace/Post/dummyData.ts
@@ -1,23 +1,0 @@
-const dummyArticles = {
-  writer: {
-    id: 1,
-    username: "홍길동",
-    email: "~~",
-    avatar: "images/man.png",
-  },
-  challenge: {
-    title: "일회용품 No! 다시 쓰기 Yes!",
-    description: "이 챌린지는 블라블라 ~",
-  },
-  created_at: "2022.08.10",
-  content:
-    "이번 챌린지는 정말 어려웠습니다. 각급 선거관리위원회는 선거인명부의 작성등 선거사무와 국민투표사무에 관하여 관계 행정기관에 필요한 지시를 할 수 있다. 헌법재판소는 법관의 자격을 가진 9인의 재판관으로 구성하며, 재판관은 대통령이 임명한다. 정당은 법률이 정하는 바에 의하여 국가의 보호를 받으며, 국가는 법률이 정하는 바에 의하여 정당운영에 필요한...",
-  images: [
-    { id: 1, image_url: "images/image3.png" },
-    { id: 2, image_url: "images/image4.png" },
-    { id: 3, image_url: "images/image5.png" },
-    { id: 4, image_url: "images/image3.png" },
-  ],
-};
-
-export default dummyArticles;
