+++ src/pages/EndedChallenge.tsx
@@ -11,7 +11,7 @@ const dummyChallenges: ChallengeType[] = [
   {
     id: "1",
     title: "텀블러로 커피 마시는 멋진 나는..",
-    desciption: "내용",
+    description: "내용",
     categories: ["001a", "텀블러_챌린지"],
     prover_cnt: 10,
     point: 100,
@@ -27,7 +27,7 @@ const dummyChallenges: ChallengeType[] = [
   {
     id: "1",
     title: "텀블러로 커피 마시는 멋진 나는..",
-    desciption: "내용",
+    description: "내용",
     categories: ["001a", "텀블러_챌린지"],
     prover_cnt: 10,
     point: 100,
@@ -43,7 +43,7 @@ const dummyChallenges: ChallengeType[] = [
   {
     id: "1",
     title: "텀블러로 커피 마시는 멋진 나는..",
-    desciption: "내용",
+    description: "내용",
     categories: ["001a", "텀블러_챌린지"],
     prover_cnt: 10,
     point: 100,
@@ -59,7 +59,7 @@ const dummyChallenges: ChallengeType[] = [
   {
     id: "1",
     title: "텀블러로 커피 마시는 멋진 나는..",
-    desciption: "내용",
+    description: "내용",
     categories: ["001a", "텀블러_챌린지"],
     prover_cnt: 10,
     point: 100,
@@ -75,7 +75,7 @@ const dummyChallenges: ChallengeType[] = [
   {
     id: "1",
     title: "텀블러로 커피 마시는 멋진 나는..",
-    desciption: "내용",
+    description: "내용",
     categories: ["001a", "텀블러_챌린지"],
     prover_cnt: 10,
     point: 100,
@@ -91,7 +91,7 @@ const dummyChallenges: ChallengeType[] = [
   {
     id: "1",
     title: "텀블러로 커피 마시는 멋진 나는..",
-    desciption: "내용",
+    description: "내용",
     categories: ["001a", "텀블러_챌린지"],
     prover_cnt: 10,
     point: 100,
@@ -107,7 +107,7 @@ const dummyChallenges: ChallengeType[] = [
   {
     id: "1",
     title: "텀블러로 커피 마시는 멋진 나는..",
-    desciption: "내용",
+    description: "내용",
     categories: ["001a", "텀블러_챌린지"],
     prover_cnt: 10,
     point: 100,
@@ -123,7 +123,7 @@ const dummyChallenges: ChallengeType[] = [
   {
     id: "1",
     title: "텀블러로 커피 마시는 멋진 나는..",
-    desciption: "내용",
+    description: "내용",
     categories: ["001a", "텀블러_챌린지"],
     prover_cnt: 10,
     point: 100,
@@ -139,7 +139,7 @@ const dummyChallenges: ChallengeType[] = [
   {
     id: "1",
     title: "텀블러로 커피 마시는 멋진 나는..",
-    desciption: "내용",
+    description: "내용",
     categories: ["001a", "텀블러_챌린지"],
     prover_cnt: 10,
     point: 100,
@@ -155,7 +155,7 @@ const dummyChallenges: ChallengeType[] = [
   {
     id: "1",
     title: "텀블러로 커피 마시는 멋진 나는..",
-    desciption: "내용",
+    description: "내용",
     categories: ["001a", "텀블러_챌린지"],
     prover_cnt: 10,
     point: 100,
@@ -171,7 +171,7 @@ const dummyChallenges: ChallengeType[] = [
   {
     id: "1",
     title: "텀블러로 커피 마시는 멋진 나는..",
-    desciption: "내용",
+    description: "내용",
     categories: ["001a", "텀블러_챌린지"],
     prover_cnt: 10,
     point: 100,
@@ -187,7 +187,7 @@ const dummyChallenges: ChallengeType[] = [
   {
     id: "1",
     title: "텀블러로 커피 마시는 멋진 나는..",
-    desciption: "내용",
+    description: "내용",
     categories: ["001a", "텀블러_챌린지"],
     prover_cnt: 10,
     point: 100,
@@ -203,7 +203,7 @@ const dummyChallenges: ChallengeType[] = [
   {
     id: "1",
     title: "텀블러로 커피 마시는 멋진 나는..",
-    desciption: "내용",
+    description: "내용",
     categories: ["001a", "텀블러_챌린지"],
     prover_cnt: 10,
     point: 100,
@@ -219,7 +219,7 @@ const dummyChallenges: ChallengeType[] = [
   {
     id: "1",
     title: "텀블러로 커피 마시는 멋진 나는..",
-    desciption: "내용",
+    description: "내용",
     categories: ["001a", "텀블러_챌린지"],
     prover_cnt: 10,
     point: 100,
@@ -235,7 +235,7 @@ const dummyChallenges: ChallengeType[] = [
   {
     id: "1",
     title: "텀블러로 커피 마시는 멋진 나는..",
-    desciption: "내용",
+    description: "내용",
     categories: ["001a", "텀블러_챌린지"],
     prover_cnt: 10,
     point: 100,
@@ -251,7 +251,7 @@ const dummyChallenges: ChallengeType[] = [
   {
     id: "1",
     title: "텀블러로 커피 마시는 멋진 나는..",
-    desciption: "내용",
+    description: "내용",
     categories: ["001a", "텀블러_챌린지"],
     prover_cnt: 10,
     point: 100,
